Update PathWalker logic to handle diagonal movements better.
We now update the direction to walk only if the next tile has changed. This means diagonal movement is better because we don't change direction until we reach the tile we need. Previously, because player location is a float, sometimes we would end up in the tile next to the tile we need and stop moving diagonally. This meant the player was partially in two tiles and obstacles in the way could cause "strafing".

This should now be fixed. However, I had to add a kill switch for if we accidentally end up completely off course (we probably needed this earlier too). This hasn't happened in the tests yet, but it's useful to have the kill switch.