Added support for TileContents to SimpleMap.
SimpleMap now generates tiles which contain accurate information about where there is a twig, stone, tree or weed on that tile. I am not sure what the performance hit for doing all of this up front is.

Note, SimpleTile and what it contains will probably evolve constantly as we figure out all the things that are relevant to us. We might get to a point where it's simpler to just have utility methods to abstract away game access and use game data directly for PathFinder.