Reduce PathFinder time taken for large grids massively.
Map creation (sample size 10):       18.9
Graph creation (sample size 10):     25
Path finding (sample size 10):       36.8

This commit adds a "visitedTiles" hashset which means we no longer need to add all the tiles to "tilesToVisit" hashset. This makes finding the closest tile much faster on large maps, with the small cost of having to check the presence of each neighbour inside a hashset but that's an O(1) operation.